refactor(search): split legacy and reworked search
* reworked is the new version behind a feature flag
* legacy will be deleted after the feature flag has been enabled
